Minor edit to CenterToCube
